organized onmachina api code
