+++ lib/nearsetup.js
@@ -5,6 +5,7 @@ import { setupNearWallet } from '@near-wallet-selector/near-wallet';
 import { setupMyNearWallet } from '@near-wallet-selector/my-near-wallet';
 import '@near-wallet-selector/modal-ui/styles.css';
 import * as nearAPI from 'near-api-js';
+import * as onMachina from './onmachina';
 import { Buffer } from 'buffer';
 
 // buffer pollyfill needed by Near Wallet Selector
@@ -49,32 +50,8 @@ console.log('publicKey: ', signatureData.publicKey.toString());
 const signature = Buffer.from(signatureData.signature);
 
 export default async function nearSetup() {
-  const x_auth_token = await refreshAuthToken();
+  const x_auth_token = await onMachina.getAuthToken(accountId);
   return { selectorWallet, account, accountId, x_auth_token };
-}
-
-export async function refreshAuthToken() {
-  console.log('requesting auth token from OnMachina...');
-  const credentials = Buffer.from(
-    JSON.stringify({
-      account: accountData.toString('base64'),
-      publicKey: publicKey.toString('base64'),
-      signature: signature.toString('base64'),
-    }),
-  ).toString('base64');
-
-  const response = await fetch('https://auth.testnet.onmachina.io/auth/v1', {
-    method: 'GET',
-    headers: {
-      'x-auth-user': accountId,
-      'x-auth-key': credentials,
-    },
-  });
-
-  // Retrieve the x-auth-token
-  const x_auth_token = response.headers.get('x-auth-token');
-  console.log('x_auth_token: ', x_auth_token);
-  return x_auth_token;
 }
 
 export async function getAccountId() {
