+++ src/contexts/NearContext.jsx
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect, useState } from 'react';
-import * as NearAccount from '../../lib/newnearsetup';
+import * as NearAccount from '../../lib/nearsetup';
 
 export const NearAccountContext = React.createContext();
